+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+#### [3.6.0(28)]
+
+API
+
+- `VCoreHumanReadableLocalizationProvider` is added, that automatically localized errors, and only exposes human-readable strings
+- `DefaultVCoreLocalizationProvider` can now be initialized
+
 #### [3.5.1(27)](https://github.com/VakhoKontridze/VCore/releases/tag/3.5.1) — *2022 06 15*
 
 Architectural Pattern Helpers